requirements.txt
flask
gunicorn
flask-cors
langchain
langchain-community
langchain-core
langchain-text-splitters
langgraph
pypdf
faiss-cpu
sentence-transformers
pydantic
python-dotenv
tiktoken
tqdm
langchain-huggingface
SpeechRecognition
langchain-groq
requests
pdfplumber
python-docx
tenacity
langsmith
